+++ JIUse/src/br/com/bhansen/iuc/metric/CheckMoves.java
@@ -3,12 +3,12 @@ package br.com.bhansen.iuc.metric;
 public class CheckMoves implements Metric{
 
 	@Override
-	public float getMetric() throws Exception {
+	public double getMetric() throws Exception {
 		return 0;
 	}
 
 	@Override
-	public float getMetric(String fakeDelegate, String fakeParameter) throws Exception {
+	public double getMetric(String fakeDelegate, String fakeParameter) throws Exception {
 		return 0;
 	}
