+++ JIUse/src/br/com/bhansen/iuc/metric/Metric.java
@@ -2,7 +2,7 @@ package br.com.bhansen.iuc.metric;
 
 public interface Metric {
 	
-	public float getMetric() throws Exception;
-	public float getMetric(String fakeDelegate, String fakeParameter) throws Exception;
+	public double getMetric() throws Exception;
+	public double getMetric(String fakeDelegate, String fakeParameter) throws Exception;
 
 }
